+++ examples/stm32f107/src/bin/eth.rs
@@ -73,7 +73,7 @@ async fn main(spawner: Spawner) -> ! {
 
     // Add the network interface to the stack.
     static DEVICE: StaticCell<Device> = StaticCell::new();
-    let iface = unwrap!(stack.add_iface(DEVICE.init(device)).ok());
+    let iface = unwrap!(stack.add_iface(DEVICE.init(device)));
     iface.set_dhcpv4(Some(Default::default()));
     spawner.spawn(unwrap!(net_task(runner)));
     iface.wait_config_up().await;
@@ -83,7 +83,7 @@ async fn main(spawner: Spawner) -> ! {
     let mut rx_buffer = [0; 1024];
     let mut tx_buffer = [0; 1024];
 
-    let mut listener = TcpListener::new(stack);
+    let mut listener = unwrap!(TcpListener::new(stack));
     unwrap!(listener.listen(IpListenEndpoint { addr: None, port: 80 }));
 
     loop {